check user authorization to edit and delete

+++ app/Http/Controllers/ImageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ImageRequest;
 use App\Models\Image;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ImageController extends Controller
 {
@@ -32,6 +33,9 @@ class ImageController extends Controller
 
     public function edit(Image $image)
     {
+        if (request()->user()->id != $image->user_id){
+            abort(403, "Access denied");
+        }
         return view('image.edit', compact('image'));
     }
 
@@ -43,6 +47,9 @@ class ImageController extends Controller
 
     public function destroy(Image $image)
     {
+        if (request()->user()->id != $image->user_id){
+            abort(403, "Access denied");
+        }
         $image->delete();
         return to_route('images.index')->with('message', 'Image has been remove successfully');
     }
